requirements.txt
lxml==5.3.1
Markdown==3.7
python-docx==1.1.2
PyYAML==6.0.2
ruamel.yaml==0.18.10
typing_extensions==4.12.2
